Remove `@backbase/foundation-ang` dependency from mocks
This change does mean that the consuming interface changes a bit, the
app.module is responsible for calling `createMocks` with the exported
data. I tried to do this inside of `createMocksProvider` but wasn't able
to yet.

Also, still need to figure out how to remove the dependency from the module
without breaking lazy loading. Lazy loading causes an issue because the
configuation is lazy loaded but the mock module is not, so the
interceptor is configured before the config is really available.